Fixed the issues plaguing the system
The encryption must be able to deal with negative numbers. It looks
like that the evaluation with finite group '-1's is not sufficient. I
have to look into mathematical basis, because this might come back and
bite me later.

Also remove some of the test code to clear things up.

Test done: The intersection of the sample sets is producing the correct set.