Remove the template parameters for eval calls
The poly::eval() function must be called with bells and whistles
template parameters because the return type do not take part in the
template parameter deduction.

Solve the issue with removing the CoefficientType parameter and replace
its usages with typename PolynomialIt::value_type. Since the
PolynomialIt is an iterator to coefficients, it is a much better
solution. The only thing with this is that the PolynomialIt must have a
value_type member type.

Change all the usages to reduce the code.