Remove all unnecessary code
The original idea behind the project is pretty sound however this isn't
the right place to do this. In preparation to move the code to a public
share to pair with the lisp version, trim the code so that it only does
what it needs to show the basic concept.

Remove all unnecessary bits and includes.