Remove the exponential feature from the ElGamal system
The original code was using the exponential ElGamal encryption system
to get make the ElGamal cryptographic method homomorphic when adding.
While the idea is sound behind this system, the ElGamal cryptosystem as
such doesn't have to contain this feature. In fact, it is easier to
verify the behaviour of the ElGamal encryption and decryption without
encoding the exponential feature into it.