Implementing Exponential ElGamal
In order to retain the features for Private Set Intersection it is
necessary to use a slightly modified version of the original ElGamal
encryption scheme. Basically it means that instead of using the plain
text directly, it uses the generator exponential of the plain text which
means that what was homomorphic to multiplication before, can be
regarded as homomorphic to addition.