Transforming the ObliviousEvaluation API
Instead of having only 3 static method on a templated class it feels
more natural if we represent the sides of the communication as distinct
objects, Client and Server.

This actors store their respective state for communicating with the
other end of the wire. This is still a fluid conecpt but should work.