Returning encrypted polynomial for the server
Refactored the code such that a test can be constructed that simplify
the process. Basically taking a plain text without any encryption, so
the results can be done away without any random number generator issues.

For composability the encryption type and the input type has become
template argument.

The Encryption concept has been extended with a Decipher function which
helps to map the results of decryption to the original elements as the
Exponential ElGamal method itself doesn't deciper straight into the
elements themselves but their g^e form. Since the client is not
interested in the actual elements but how they match, it easier and
possibly faster to keep track of the g^e form to match the sets.