Use normal, signed integer as input and poly type
Using CyclicRing polynomial coefficients was a mistake. The assumption
made with Exponential El Gamal is that we can hand in normal, signed
integers because it will be used only through the generator anyway and
the result therefore will be an element of the Cyclic Ring anyway.

But for the algebra to work, the exponentiation must use signed integers
when it comes to polynomials. At the moment the polynomial type is
matched with the input type but this might be to bold of an assumption.
A set of unsigned integers should work as well as anything as long as we
can turn them into a meaningful polynomial.